third_party/WebKit/LayoutTests/editing/selection/selection-change-in-blur-event-by-remove-children.html

<div id="container">
<p id="description"></p>
<div id="sample"><span contenteditable="true">foobar</span></div>
</div>
<div id="console"></div>
<script src="../../resources/js-test.js"></script>
<script>
description('Selection should be updated after set innerHTML.');

function $(id) { return document.getElementById(id); }

var selection = window.getSelection();
document.addEventListener('blur', function () {
    selection.selectAllChildren($('sample').firstChild);
}, true);

selection.selectAllChildren($('sample').firstChild);
$('sample').firstChild.focus();
$('sample').innerHTML = '';

shouldBe('selection.anchorOffset', '0');
shouldBe('selection.focusOffset', '0');

if (window.testRunner)
    $('container').outerHTML = '';
</script>
